Updating load data process

+++ integration-py-matlab/scripts_py/load_data.py
@@ -9,20 +9,6 @@ locale.setlocale(locale.LC_TIME, 'pt_BR.utf8')
 
 def format_files(active, dt_start, dt_end, interval):
     
-    if active == 'IBOVESPA':
-        df = web.get_data_yahoo('^BVSP', start=dt_start, end=dt_end,interval=interval).reset_index()
-    else:
-        df = web.get_data_yahoo(f'{active}.SA', start=dt_start, end=dt_end,interval=interval).reset_index()
-    
-    df.rename(columns={'Date': 'data',
-                       'High': 'maxima', 
-                       'Low': 'minima', 
-                       'Open': 'abertura', 
-                       'Close': 'fechamento', 
-                       'Volume': 'volume'}, inplace=True)
-    df['variacao'] = (df['fechamento']-df['abertura'])/df['abertura']
-    df = df[['data', 'maxima', 'minima', 'abertura', 'fechamento', 'variacao']]
-    
     start_month = datetime.datetime.strptime(dt_start, "%Y-%m-%d").strftime("%b").upper()
     start_year = datetime.datetime.strptime(dt_start, "%Y-%m-%d").strftime("%y")
 
@@ -36,10 +22,30 @@ def format_files(active, dt_start, dt_end, interval):
     if not os.path.exists(outdir):
         os.mkdir(outdir)
 
-    fullname = os.path.join(outdir, f"{file_name}.csv")    
+    fullname = os.path.join(outdir, f"{file_name}.csv")
 
-    df.to_csv(fullname,index=False)
+    file_exists = os.path.exists(fullname)
 
+    if not file_exists:
+        if active == 'IBOVESPA':
+            df = web.get_data_yahoo('^BVSP', start=dt_start, end=dt_end,interval=interval).reset_index()
+        else:
+            df = web.get_data_yahoo(f'{active}.SA', start=dt_start, end=dt_end,interval=interval).reset_index()
+
+        df.rename(columns={'Date': 'data',
+                        'High': 'maxima', 
+                        'Low': 'minima', 
+                        'Open': 'abertura', 
+                        'Close': 'fechamento', 
+                        'Volume': 'volume'}, inplace=True)
+        df['variacao'] = (df['fechamento']-df['abertura'])/df['abertura']
+        df = df[['data', 'maxima', 'minima', 'abertura', 'fechamento', 'variacao']]
+
+        df.to_csv(fullname,index=False)
+
+    else:
+        df = pd.read_csv(fullname)
+    
     return df
 
 def load_data(ativo, start_date, end_date):
@@ -52,23 +58,3 @@ def load_data(ativo, start_date, end_date):
         pass
     
     return df
-
-
-# def load_data(benchmark, ativos, start_date, end_date):
-#     interval = 'm'
-#     # format_files(benchmark,start_date,end_date,interval)
-#     # for ativo in ativos:
-#     #     format_files(ativo,start_date,end_date,interval)
-#     #     print(ativo)
-#     try: 
-#         format_files(benchmark,start_date,end_date,interval)
-#         for ativo in ativos:
-#             format_files(ativo,start_date,end_date,interval)
-
-#     except Exception as e:
-#         print(e)
-#         return 1
-
-#     return 0
-
-
